typos: Perform most checks using a hashtable lookup instead of a regular expression match.
This gives us approximately a 40x speed boost. This is most important and effective now that we use various third-party common misspelling lists.
The expand_regexp() function is responsible for turning as many as possible of our internal checks into simple word lookups that can be done through the hashtable. Regular expressions that are too complex for expansion are still processed normally.